test: extract MockIssueExtractionService.swift from IssueExportTestSupport.swift
Splits the `MockIssueExtractionService` (IssueExtracting double) out
of IssueExportTestSupport.swift into its own file so that support file
holds only the primary `MockExportService`.

Byte-identical move of lines 4-19. Both files retain their test imports.

IssueExportTestSupport.swift: 362 → 344 lines. Tests: 667.
