+++ Tests/BugNarratorTests/IssueExportTestSupport.swift
@@ -1,23 +1,5 @@
 import Foundation
 @testable import BugNarrator
-
-actor MockIssueExtractionService: IssueExtracting {
-    var result = IssueExtractionResult(summary: "", issues: [])
-
-    func setResult(_ result: IssueExtractionResult) {
-        self.result = result
-    }
-
-    func extractIssues(
-        from reviewSession: TranscriptSession,
-        apiKey: String,
-        model: String,
-        apiBaseURL: URL
-    ) async throws -> IssueExtractionResult {
-        result
-    }
-}
-
 actor MockExportService: IssueExporting {
     var gitHubResults: [ExportResult] = []
     var jiraResults: [ExportResult] = []
